Switched Todo repository to SQLite and updated related files
This commit represents a significant shift in the management of Todo items, moving from in-memory, REST, and file-based repositories to a SQLite database. The `MauiProgram.cs` file was updated to use the new `DbTodoRepository` implementation. SQLite attributes were added to the `Todo.cs` file to facilitate database operations. Constants related to the SQLite database were added to the `Constants.cs` file. Thread safety was improved in the `FileTodoRepository.cs` file using a `SemaphoreSlim` object. The `ITodoRepository.cs` file was updated to reflect changes in the return type of the `GetTodoById` method. A new observable property for the `Id` of the Todo item was added to the `EditTodoViewModel.cs` file. The `Week3Lab.csproj` file was updated to include package references for SQLite libraries. A new file, `DbTodoRepository.cs`, was added to implement the `ITodoRepository` interface using a SQLite database.